rmtree fixed

+++ src/plugins/intel_cpu/tools/commit_slider/utils/helpers.py
@@ -81,13 +81,15 @@ def absolutizePaths(cfg):
     pl = sys.platform
     if pl == "linux" or pl == "linux2":
         cfg["workPath"] = cfg["linWorkPath"]
+        cfg["os"] = "linux"
     elif pl == "win32":
         wp = cfg["winWorkPath"]
         wp = "echo {path}".format(path=wp)
         wp = subprocess.check_output(wp, shell=True)
         wp = wp.decode()
         wp = wp.rstrip()
-        cfg["workPath"] = wp
+        cfg["workPath"] = wpy
+        cfg["os"] = "win"
     else:
         raise CfgError(
             "No support for current OS: {pl}".format(pl=pl)
@@ -273,12 +275,12 @@ def getActualPath(pathName, cfg):
     return curPath.format(workPath=workPath)
 
 
-def safeClearDir(path):
+def safeClearDir(path, cfg):
     if not os.path.exists(path):
         os.makedirs(path)
-    try:
+    if cfg["os"] == "win":
         shutil.rmtree(path)
-    except PermissionError:
+    else:
         # WA, because of unstability of rmtree()
         # in linux environment
         p = subprocess.Popen(
